requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
openai
python-dotenv
requests
httpx
alembic
aiosqlite
